pyproject.toml
[project]
name = "viral-reels-agent"
version = "0.1.0"
description = "Viral Reels discovery, approval and auto-publish automation for Instagram & YouTube Shorts"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "apify-client>=1.7",
    "instagrapi>=2.1",
    "google-api-python-client>=2.130",
    "google-auth-oauthlib>=1.2",
    "yt-dlp>=2024.8.6",
    "videohash>=3.0.1",
    "ffmpeg-python>=0.2.0",
    "APScheduler>=3.10",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "pydantic-settings>=2.3",
    "httpx>=0.27",
    "loguru>=0.7",
    "claude-agent-sdk>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
